+++ random_profile/test_package.py
@@ -5,20 +5,12 @@ random_profile = RandomProfile(num=1)
 
 
 class RandomProfileTest(unittest.TestCase):
-    def test_fname(self):
-        self.assertEqual(len(random_profile.first_name()), 1)
-
     def test_faname_with_num(self):
         self.assertEqual(len(RandomProfile(num=10).first_name()), 10)
-
-    def test_lname(self):
-        self.assertEqual(len(random_profile.last_name()), 1)
 
     def test_lname_with_num(self):
         self.assertEqual(len(RandomProfile(num=10).last_name()), 10)
 
-    def test_full_name(self):
-        self.assertEqual(len(random_profile.full_name()), 1)
 
     def test_full_name_with_num(self):
         self.assertEqual(len(RandomProfile(num=10).full_name()), 10)